Remove Current Queue Tab, replace with a skeleton (spooky!)
So there's no way it seems to manipulate/get data from the current queue via the Spotify API, so my hopes of using that are dashed. So removing the tab that would've used that. I left in it's place a skeleton, per se, that I'll use for a new tab.

I got some ideas... yeahhhh